refactor: split agent.rs into sub-modules (compact, spawner)
- src/agent/mod.rs: Agent struct, core methods, send/execute internals, tests (3341 lines)
- src/agent/compact.rs: compact() method and prompt builder (118 lines)
- src/agent/spawner.rs: HeadlessOutcome, run_headless, AgentSpawner, spawn helpers (256 lines)
