+++ src/agent/mod.rs
@@ -2,7 +2,7 @@ use std::sync::{Arc, Mutex};
 use tokio::sync::Mutex as TokioMutex;
 use tokio_util::sync::CancellationToken;
 
-use crate::backend::{BackendFactory, LlmBackend};
+use crate::backend::LlmBackend;
 use crate::config::{ConfirmationMode, ToolsConfig};
 use crate::context_files::{ContextFile, discover_context_files_from_env};
 use crate::session::Session;
@@ -15,6 +15,11 @@ use anyhow::Result;
 use futures::StreamExt;
 use futures::channel::mpsc;
 use futures::future::join_all;
+
+mod compact;
+mod spawner;
+
+pub use spawner::*;
 
 struct PendingToolCall {
     id: String,
@@ -174,116 +179,6 @@ impl Agent {
             .conversation()
             .load_history()
             .await
-    }
-
-    /// Run context compaction: summarise the active conversation history using a
-    /// headless sub-agent, then replace it with the summary while preserving the
-    /// context prefix (skills, CLAUDE.md, etc.).
-    ///
-    /// Returns the summary text on success, or an error message on failure.
-    pub async fn compact(&self) -> Result<String, String> {
-        let spawner = match &self.compaction_spawner {
-            Some(s) => Arc::clone(s),
-            None => {
-                return Err("Compaction not available: no spawner configured.".to_string());
-            }
-        };
-
-        let compaction_role = spawner.app_config.compaction_role.clone();
-
-        let role = if spawner.app_config.models.contains_key(&compaction_role) {
-            compaction_role
-        } else {
-            "default".to_string()
-        };
-
-        let prompt = {
-            let history = lock(&self.history);
-            let prefix_len = *self
-                .context_prefix_len
-                .lock()
-                .unwrap_or_else(|e| e.into_inner());
-            let persisted = &history[prefix_len.min(history.len())..];
-            let mut parts = Vec::new();
-            for msg in persisted {
-                let role_label = match msg.role {
-                    Role::User => "User",
-                    Role::Assistant => "Assistant",
-                };
-                for block in &msg.content {
-                    match block {
-                        ContentBlock::Text(text) => {
-                            parts.push(format!("{role_label}: {text}"));
-                        }
-                        ContentBlock::ToolUse { name, input, .. } => {
-                            parts.push(format!(
-                                "{role_label}: [Called tool {name} with input {input}]"
-                            ));
-                        }
-                        ContentBlock::ToolResult {
-                            content, is_error, ..
-                        } => {
-                            let label = if *is_error { "error" } else { "result" };
-                            parts.push(format!("{role_label}: [Tool {label}: {content}]"));
-                        }
-                        ContentBlock::Thinking { text, .. } => {
-                            parts.push(format!("{role_label}: [Thinking: {text}]"));
-                        }
-                        ContentBlock::RedactedThinking { .. } => {
-                            parts.push(format!("{role_label}: [Redacted thinking]"));
-                        }
-                    }
-                }
-            }
-            if parts.is_empty() {
-                return Ok("Nothing to compact: the conversation is empty.".to_string());
-            }
-            format!(
-                "Summarize the following conversation concisely, preserving key facts, decisions, and context that would be needed to continue the conversation. Do not include meta-commentary — output only the summary.\n\n{}",
-                parts.join("\n\n")
-            )
-        };
-
-        let outcome = spawner
-            .spawn(&role, ConfirmationMode::Never, Some(&[]), prompt)
-            .await;
-
-        if outcome.is_error {
-            let msg = outcome
-                .error_message
-                .unwrap_or_else(|| "Compaction failed with an unknown error.".to_string());
-            return Err(msg);
-        }
-
-        let summary = outcome.text;
-        if summary.trim().is_empty() {
-            return Err("Compaction produced an empty summary.".to_string());
-        }
-
-        let summary_msg = Message::text(Role::User, format!("[Compacted] {summary}"));
-
-        let session = self.session.lock().await;
-
-        if let Err(e) = session.conversation().compact(&summary_msg).await {
-            return Err(format!("Compaction failed: {e}"));
-        }
-
-        drop(session);
-
-        let prefix_len = *self
-            .context_prefix_len
-            .lock()
-            .unwrap_or_else(|e| e.into_inner());
-        let prefix: Vec<Message> = {
-            let history = lock(&self.history);
-            let take = prefix_len.min(history.len());
-            history[..take].to_vec()
-        };
-        lock(&self.history).clear();
-        lock(&self.history).extend(prefix);
-        lock(&self.history).push(summary_msg.clone());
-
-        Ok(summary)
     }
 
     /// Return a snapshot of all tasks in the current session.
@@ -946,250 +841,6 @@ async fn execute_tool_calls(
 }
 
 pub(crate) const DEFAULT_MAX_TOKENS: u32 = 8_192;
-
-/// Outcome of a headless sub-agent run.
-pub struct HeadlessOutcome {
-    pub text: String,
-    pub input_tokens: u32,
-    pub output_tokens: u32,
-    pub is_error: bool,
-    pub error_message: Option<String>,
-}
-
-/// Drive an `Agent` to completion without a human in the loop.
-///
-/// Any `ToolConfirmationRequired` event causes an immediate error — sub-agents
-/// must be configured with a confirmation mode that does not require human input.
-pub async fn run_headless(agent: &Agent, prompt: String) -> HeadlessOutcome {
-    let stream = match agent.send(prompt, None, None).await {
-        Ok(s) => s,
-        Err(e) => {
-            return HeadlessOutcome {
-                text: String::new(),
-                input_tokens: 0,
-                output_tokens: 0,
-                is_error: true,
-                error_message: Some(e.to_string()),
-            };
-        }
-    };
-
-    let mut stream = stream;
-    let mut text = String::new();
-    let mut input_tokens: u32 = 0;
-    let mut output_tokens: u32 = 0;
-    let mut is_error = false;
-    let mut error_message: Option<String> = None;
-
-    while let Some(event) = stream.next().await {
-        match event {
-            AgentEvent::TokenReceived(t) => text.push_str(&t),
-            AgentEvent::ResponseComplete(_) => {}
-            AgentEvent::ToolUseReceived { .. } => {}
-            AgentEvent::ToolResult { .. } => {}
-            AgentEvent::ToolConfirmationRequired { name, .. } => {
-                is_error = true;
-                error_message = Some(format!(
-                    "Sub-agent required confirmation for tool '{name}' but no human is present. \
-                     Set a less restrictive confirmation mode for the sub-agent."
-                ));
-                break;
-            }
-            AgentEvent::Error(msg) => {
-                is_error = true;
-                error_message = Some(msg);
-                break;
-            }
-            AgentEvent::Usage {
-                input_tokens: it,
-                output_tokens: ot,
-                ..
-            } => {
-                input_tokens = input_tokens.saturating_add(it);
-                output_tokens = output_tokens.saturating_add(ot);
-            }
-            AgentEvent::SubAgentUsage { .. } => {}
-            AgentEvent::Interrupted { .. } => {}
-            AgentEvent::Warn(_) => {}
-            AgentEvent::ThinkingReceived(_) => {}
-            AgentEvent::CompactionComplete { .. } => {}
-        }
-    }
-
-    HeadlessOutcome {
-        text,
-        input_tokens,
-        output_tokens,
-        is_error,
-        error_message,
-    }
-}
-
-/// Returns the stricter of two confirmation modes.
-///
-/// Strictness ordering: `Always` > `WriteOnly` > `Never`.
-/// The sub-agent can never be more permissive than the parent.
-pub fn clamp_confirmation(
-    parent: &ConfirmationMode,
-    requested: Option<&ConfirmationMode>,
-) -> ConfirmationMode {
-    let requested = match requested {
-        Some(r) => r,
-        None => return parent.clone(),
-    };
-
-    match (parent, requested) {
-        (ConfirmationMode::Always, _) => ConfirmationMode::Always,
-        (ConfirmationMode::WriteOnly, ConfirmationMode::Always) => ConfirmationMode::Always,
-        (ConfirmationMode::WriteOnly, _) => ConfirmationMode::WriteOnly,
-        (ConfirmationMode::Never, ConfirmationMode::Always) => ConfirmationMode::Always,
-        (ConfirmationMode::Never, ConfirmationMode::WriteOnly) => ConfirmationMode::WriteOnly,
-        (ConfirmationMode::Never, ConfirmationMode::Never) => ConfirmationMode::Never,
-    }
-}
-
-/// Builds a fresh `ToolRegistry` for a sub-agent.
-///
-/// The closure receives the sub-agent's session so that session-bound tools
-/// (e.g. task tools) are wired to the sub-agent rather than the parent.
-pub type RegistryBuilder =
-    Box<dyn Fn(Arc<tokio::sync::Mutex<Session>>) -> anyhow::Result<ToolRegistry> + Send + Sync>;
-
-/// Factory used by `AgentTool` to spawn independent sub-agents.
-pub struct AgentSpawner {
-    pub factory: Arc<BackendFactory>,
-    pub app_config: Arc<crate::config::AppConfig>,
-    pub registry_builder: RegistryBuilder,
-    pub parent_confirmation: ConfirmationMode,
-    pub skills: std::collections::HashMap<String, std::path::PathBuf>,
-}
-
-impl AgentSpawner {
-    pub async fn spawn(
-        &self,
-        role: &str,
-        confirmation: ConfirmationMode,
-        tool_allowlist: Option<&[String]>,
-        prompt: String,
-    ) -> HeadlessOutcome {
-        let session = match Session::new(None, self.app_config.sessions_dir.clone()).await {
-            Ok(s) => Arc::new(tokio::sync::Mutex::new(s)),
-            Err(e) => {
-                return HeadlessOutcome {
-                    text: String::new(),
-                    input_tokens: 0,
-                    output_tokens: 0,
-                    is_error: true,
-                    error_message: Some(format!("Failed to create sub-agent session: {e}")),
-                };
-            }
-        };
-
-        let registry = match (self.registry_builder)(Arc::clone(&session)) {
-            Ok(r) => r,
-            Err(e) => {
-                return HeadlessOutcome {
-                    text: String::new(),
-                    input_tokens: 0,
-                    output_tokens: 0,
-                    is_error: true,
-                    error_message: Some(format!("Failed to build sub-agent registry: {e}")),
-                };
-            }
-        };
-
-        let registry = if let Some(allowlist) = tool_allowlist {
-            registry.into_filtered(allowlist)
-        } else {
-            registry
-        };
-
-        let tool_config = ToolsConfig {
-            confirmation,
-            ..self.app_config.tools.clone()
-        };
-
-        let selection = match self.factory.for_role(role).await {
-            Ok(s) => s,
-            Err(e) => {
-                return HeadlessOutcome {
-                    text: String::new(),
-                    input_tokens: 0,
-                    output_tokens: 0,
-                    is_error: true,
-                    error_message: Some(format!("Failed to resolve role '{role}': {e}")),
-                };
-            }
-        };
-
-        let agent =
-            match spawn_agent_with_selection(selection, &tool_config, session, registry).await {
-                Ok(a) => a.with_thinking(self.app_config.thinking.clone()),
-                Err(e) => {
-                    return HeadlessOutcome {
-                        text: String::new(),
-                        input_tokens: 0,
-                        output_tokens: 0,
-                        is_error: true,
-                        error_message: Some(format!("Failed to spawn sub-agent: {e}")),
-                    };
-                }
-            };
-
-        let agent = match agent.with_context_files() {
-            Ok(a) => a,
-            Err(e) => {
-                return HeadlessOutcome {
-                    text: String::new(),
-                    input_tokens: 0,
-                    output_tokens: 0,
-                    is_error: true,
-                    error_message: Some(format!("Failed to load context files: {e}")),
-                };
-            }
-        };
-
-        let agent = agent.with_skills(&self.skills);
-        run_headless(&agent, prompt).await
-    }
-}
-
-/// Spawn a fresh `Agent` for the given named role using the shared `BackendFactory`.
-///
-/// The caller supplies a pre-built `ToolRegistry` and a `Session`. The agent's
-/// model is taken from the role definition; `tool_config` controls iteration
-/// limits and confirmation behaviour.
-pub async fn spawn_agent(
-    factory: &BackendFactory,
-    role: &str,
-    tool_config: &ToolsConfig,
-    session: Arc<TokioMutex<Session>>,
-    tools: ToolRegistry,
-) -> anyhow::Result<Agent> {
-    let selection = factory.for_role(role).await?;
-    spawn_agent_with_selection(selection, tool_config, session, tools).await
-}
-
-/// Core of `spawn_agent` — constructs an `Agent` from an already-resolved
-/// `BackendSelection`. Separated out so tests can inject a fake backend without
-/// going through real auth.
-pub async fn spawn_agent_with_selection(
-    selection: crate::backend::BackendSelection,
-    tool_config: &ToolsConfig,
-    session: Arc<TokioMutex<Session>>,
-    tools: ToolRegistry,
-) -> anyhow::Result<Agent> {
-    let request_config = RequestConfig {
-        model: selection.model,
-        max_tokens: DEFAULT_MAX_TOKENS,
-        tools: tools.definitions(),
-        thinking: None,
-    };
-    Ok(Agent::new(selection.backend, request_config, session)
-        .await
-        .with_tools(tools)
-        .with_tool_config(tool_config))
-}
 
 #[cfg(test)]
 mod tests {
